+++ .github/workflows/update-server.yml
@@ -79,6 +79,6 @@ jobs:
           context: .
           file: Dockerfile
           push: true
-#          tags: beclab/files-server:${{ github.event.inputs.tags }}
-          tags: beclab/files-server:${{ steps.get-latest-tag.outputs.tag }}
+          tags: beclab/files-server:${{ github.event.inputs.tags }}
+#          tags: beclab/files-server:${{ steps.get-latest-tag.outputs.tag }}
           platforms: linux/amd64,linux/arm64
