--- pkg/files/preview.go
@@ -1,1 +1,0 @@
-package files
