+++ pkg/http/upload.go
@@ -50,17 +50,6 @@ func uploadLinkHandler(w http.ResponseWriter, r *http.Request, d *common.Data) (
 		return seahub.HandleUploadLink(w, r, d)
 	}
 
-	//if strings.HasPrefix(path, upload.CacheRequestPrefix) {
-	//	path = upload.CachePathPrefix + strings.TrimPrefix(path, upload.CacheRequestPrefix)
-	//	path = upload.RewriteUrl(path, cachePvc, upload.CachePathPrefix, "")
-	//} else {
-	//	focusPrefix := "/"
-	//	if strings.HasPrefix(path, "/data/") {
-	//		focusPrefix = "/data/"
-	//	}
-	//	path = upload.RewriteUrl(path, userPvc, "", focusPrefix)
-	//}
-
 	// change temp file location
 	extracted := upload.ExtractPart(path)
 	if extracted != "" {
@@ -117,17 +106,6 @@ func uploadedBytesHandler(w http.ResponseWriter, r *http.Request, d *common.Data
 	if fileParam.FileType == "sync" {
 		return seahub.HandleUploadedBytes(w, r, d)
 	}
-
-	//if strings.HasPrefix(parentDir, upload.CacheRequestPrefix) {
-	//	parentDir = upload.CachePathPrefix + strings.TrimPrefix(parentDir, upload.CacheRequestPrefix)
-	//	parentDir = upload.RewriteUrl(parentDir, cachePvc, upload.CachePathPrefix, "")
-	//} else {
-	//	focusPrefix := "/"
-	//	if strings.HasPrefix(parentDir, "/data/") {
-	//		focusPrefix = "/data/"
-	//	}
-	//	parentDir = upload.RewriteUrl(parentDir, userPvc, "", focusPrefix)
-	//}
 
 	if !upload.CheckDirExist(parentDir) {
 		klog.Warningf("Storage path %s is not exist or is not a dir", parentDir)
@@ -231,21 +209,6 @@ func uploadChunksHandler(w http.ResponseWriter, r *http.Request, d *common.Data)
 		klog.Warningf("uploadID:%s, err:%v", uploadID, err)
 		return http.StatusBadRequest, e.New("param invalid")
 	}
-
-	//parentDir := resumableInfo.ParentDir
-	//if !strings.HasSuffix(parentDir, "/") {
-	//	parentDir += "/"
-	//}
-	//if strings.HasPrefix(parentDir, upload.CacheRequestPrefix) {
-	//	parentDir = upload.CachePathPrefix + strings.TrimPrefix(parentDir, upload.CacheRequestPrefix)
-	//	parentDir = upload.RewriteUrl(parentDir, cachePvc, upload.CachePathPrefix, "")
-	//} else {
-	//	focusPrefix := "/"
-	//	if strings.HasPrefix(parentDir, "/data/") {
-	//		focusPrefix = "/data/"
-	//	}
-	//	parentDir = upload.RewriteUrl(parentDir, userPvc, "", focusPrefix)
-	//}
 
 	p := resumableInfo.ParentDir
 	if !strings.HasSuffix(p, "/") {
